% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixomics.R
\name{mixomics_biplot}
\alias{mixomics_biplot}
\title{Create biplot directly from mixOmics object}
\usage{
mixomics_biplot(
  mixomics_object,
  x_component = 1,
  y_component = 2,
  color_by = NULL,
  ...
)
}
\arguments{
\item{mixomics_object}{A mixOmics object (PCA, PLS, sPLS, sPLS-DA, etc.)}

\item{x_component}{X-axis component number. Default is 1.}

\item{y_component}{Y-axis component number. Default is 2.}

\item{color_by}{Optional variable to color points by.}

\item{...}{Additional arguments passed to the underlying plot function}
}
\value{
A ggplot2 object
}
\description{
Create a biplot directly from a mixOmics object.
}
\examples{
\dontrun{
library(mixOmics)
data(wine)
pca_mix <- pca(wine$X, ncomp = 3)
mixomics_biplot(pca_mix)
}

}
