% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixomics.R
\name{mixomics_loadings_plot}
\alias{mixomics_loadings_plot}
\title{Create loadings plot directly from mixOmics object}
\usage{
mixomics_loadings_plot(mixomics_object, x_component = 1, y_component = 2, ...)
}
\arguments{
\item{mixomics_object}{A mixOmics object (PCA, PLS, sPLS, sPLS-DA, etc.)}

\item{x_component}{X-axis component number. Default is 1.}

\item{y_component}{Y-axis component number. Default is 2.}

\item{...}{Additional arguments passed to the underlying plot function}
}
\value{
A ggplot2 object
}
\description{
Create a ggplot2 loadings plot directly from a mixOmics object.
}
\examples{
\dontrun{
library(mixOmics)
data(wine)
pca_mix <- pca(wine$X, ncomp = 3)
mixomics_loadings_plot(pca_mix)
}

}
