% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixomics.R
\name{mixomics_pls_to_ggmultivar}
\alias{mixomics_pls_to_ggmultivar}
\title{Extract data from mixOmics PLS object}
\usage{
mixomics_pls_to_ggmultivar(mixomics_pls, n_components = 2)
}
\arguments{
\item{mixomics_pls}{A PLS object from mixOmics::pls()}

\item{n_components}{Number of components to extract. Default is 2.}
}
\value{
A list in the same format as perform_pls()
}
\description{
Convert a mixOmics PLS object to ggmultivar format.
}
\examples{
\dontrun{
library(mixOmics)
data(wine)
pls_mix <- pls(wine$X, wine$Y, ncomp = 3)
pls_result <- mixomics_pls_to_ggmultivar(pls_mix)
pls_scores_plot(pls_result)
}

}
