% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixomics.R
\name{mixomics_spls_to_ggmultivar}
\alias{mixomics_spls_to_ggmultivar}
\title{Extract data from mixOmics sPLS object}
\usage{
mixomics_spls_to_ggmultivar(mixomics_spls, n_components = 2)
}
\arguments{
\item{mixomics_spls}{A sPLS object from mixOmics::spls()}

\item{n_components}{Number of components to extract. Default is 2.}
}
\value{
A list in the same format as perform_pls()
}
\description{
Convert a mixOmics sPLS (sparse PLS) object to ggmultivar format.
}
\examples{
\dontrun{
library(mixOmics)
data(wine)
spls_mix <- spls(wine$X, wine$Y, ncomp = 3)
spls_result <- mixomics_spls_to_ggmultivar(spls_mix)
pls_scores_plot(spls_result)
}

}
