% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixomics.R
\name{mixomics_splsda_to_ggmultivar}
\alias{mixomics_splsda_to_ggmultivar}
\title{Extract data from mixOmics sPLS-DA object}
\usage{
mixomics_splsda_to_ggmultivar(mixomics_splsda, n_components = 2)
}
\arguments{
\item{mixomics_splsda}{A sPLS-DA object from mixOmics::splsda()}

\item{n_components}{Number of components to extract. Default is 2.}
}
\value{
A list in the same format as perform_pls() with additional DA-specific information
}
\description{
Convert a mixOmics sPLS-DA object to ggmultivar format.
}
\examples{
\dontrun{
library(mixOmics)
data(wine)
splsda_mix <- splsda(wine$X, wine$Y, ncomp = 3)
splsda_result <- mixomics_splsda_to_ggmultivar(splsda_mix)
pls_scores_plot(splsda_result, color_by = wine$Y)
}

}
