% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixomics.R
\name{mixomics_to_ggmultivar}
\alias{mixomics_to_ggmultivar}
\title{Extract data from mixOmics PCA object}
\usage{
mixomics_to_ggmultivar(mixomics_object, n_components = 2)

mixomics_to_ggmultivar(mixomics_object, n_components = 2)
}
\arguments{
\item{mixomics_object}{A mixOmics object (PCA, PLS, sPLS, sPLS-DA, etc.)}

\item{n_components}{Number of components to extract. Default is 2.}

\item{mixomics_pca}{A PCA object from mixOmics::pca()}
}
\value{
A list in the same format as perform_pca()

A list in ggmultivar format
}
\description{
Convert a mixOmics PCA object to ggmultivar format.

Automatically detect and convert mixOmics objects to ggmultivar format.
}
\examples{
\dontrun{
library(mixOmics)
data(wine)
pca_mix <- pca(wine$X, ncomp = 3)
pca_result <- mixomics_to_ggmultivar(pca_mix)
pca_scores_plot(pca_result)
}

\dontrun{
library(mixOmics)
data(wine)

# Works with PCA
pca_mix <- pca(wine$X, ncomp = 3)
pca_result <- mixomics_to_ggmultivar(pca_mix)

# Works with PLS
pls_mix <- pls(wine$X, wine$Y, ncomp = 3)
pls_result <- mixomics_to_ggmultivar(pls_mix)

# Works with sPLS-DA
splsda_mix <- splsda(wine$X, wine$Y, ncomp = 3)
splsda_result <- mixomics_to_ggmultivar(splsda_mix)
}

}
