% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_loadings_plot}
\alias{pca_loadings_plot}
\title{Create PCA Loadings Plot}
\usage{
pca_loadings_plot(
  pca_result,
  x_component = 1,
  y_component = 2,
  show_labels = TRUE,
  label_size = 4,
  show_arrows = TRUE,
  arrow_length = 1,
  color_by = NULL,
  circle_radius = 1,
  show_circle = TRUE,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  theme = ggplot2::theme_minimal(),
  ...
)
}
\arguments{
\item{pca_result}{Result from perform_pca() function}

\item{x_component}{X-axis component number. Default is 1.}

\item{y_component}{Y-axis component number. Default is 2.}

\item{show_labels}{Logical, whether to show variable labels. Default is TRUE.}

\item{label_size}{Size of variable labels. Default is 4.}

\item{show_arrows}{Logical, whether to show arrows from origin to loadings. Default is TRUE.}

\item{arrow_length}{Scaling factor for arrow length. Default is 1.}

\item{color_by}{Optional variable to color loadings by. Can be a vector of the same length as variables.}

\item{circle_radius}{Radius of the unit circle. Default is 1.}

\item{show_circle}{Logical, whether to show unit circle. Default is TRUE.}

\item{title}{Character vector of length 2 for axis titles. Default is NULL.}

\item{xlab, ylab}{Custom axis labels. If NULL, uses default.}

\item{theme}{ggplot2 theme to apply. Default is theme_minimal().}

\item{...}{Additional arguments passed to geom_text() or geom_segment()}
}
\value{
A ggplot2 object
}
\description{
Create a ggplot2 loadings plot from PCA results.
}
\examples{
data(mtcars)
pca_result <- perform_pca(mtcars[, 1:7])
pca_loadings_plot(pca_result)

# Color loadings by absolute value
pca_loadings_plot(pca_result, color_by = abs(pca_result$loadings$loading))

}
