% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca_scores_plot}
\alias{pca_scores_plot}
\title{Create PCA Scores Plot}
\usage{
pca_scores_plot(
  pca_result,
  x_component = 1,
  y_component = 2,
  color_by = NULL,
  shape_by = NULL,
  size_by = NULL,
  show_ellipse = FALSE,
  ellipse_level = 0.95,
  show_labels = FALSE,
  label_size = 3,
  title = NULL,
  xlab = NULL,
  ylab = NULL,
  theme = ggplot2::theme_minimal(),
  ...
)
}
\arguments{
\item{pca_result}{Result from perform_pca() function}

\item{x_component}{X-axis component number. Default is 1.}

\item{y_component}{Y-axis component number. Default is 2.}

\item{color_by}{Optional variable to color points by. Can be a vector of the same length as samples.}

\item{shape_by}{Optional variable to shape points by. Can be a vector of the same length as samples.}

\item{size_by}{Optional variable to size points by. Can be a vector of the same length as samples.}

\item{show_ellipse}{Logical, whether to show confidence ellipses. Default is FALSE.}

\item{ellipse_level}{Confidence level for ellipses (0-1). Default is 0.95.}

\item{show_labels}{Logical, whether to show sample labels. Default is FALSE.}

\item{label_size}{Size of sample labels. Default is 3.}

\item{title}{Character vector of length 2 for axis titles. Default is c("PC1", "PC2").}

\item{xlab, ylab}{Custom axis labels. If NULL, uses default.}

\item{theme}{ggplot2 theme to apply. Default is theme_minimal().}

\item{...}{Additional arguments passed to geom_point()}
}
\value{
A ggplot2 object
}
\description{
Create a ggplot2 scores plot from PCA results.
}
\examples{
data(mtcars)
pca_result <- perform_pca(mtcars[, 1:7])
pca_scores_plot(pca_result) +
  ggtitle("PCA Scores Plot")

# Color by cylinder
pca_scores_plot(pca_result, color_by = mtcars$cyl)

}
